add number of connected monitors rule, if the number is equal 1, the script execution is aborted

+++ src/ubuntu-wallpaper-fix.sh
@@ -1,12 +1,20 @@
 #!/bin/sh
 running=true
+numberOfConnectedMonitors=`xrandr -q | grep ' connected' | wc -l`
 while $running; do
-  if pgrep -f "gvfsd-trash" > /dev/null; then
-      sleep 5
-      rm -rf /home/$USER/.cache/wallpaper/*
-      gsettings set org.gnome.desktop.background picture-options stretched
-      gsettings set org.gnome.desktop.background picture-options zoom
-      exit 0
+  if [ "$numberOfConnectedMonitors" -ge "2" ]; then
+    printf 'Executing...'
+    if pgrep -f "gvfsd-trash" > /dev/null; then
+        sleep 5
+        rm -rf /home/$USER/.cache/wallpaper/*
+        gsettings set org.gnome.desktop.background picture-options stretched
+        gsettings set org.gnome.desktop.background picture-options zoom
+        printf '\nDone\n'
+        exit 0
+    fi
+  else
+    printf 'Just only one monitor is connected... Nothing to do here! Beep.\n'
+    exit 0
   fi
   sleep 1
 done
